se modificaron las migraciones y se actualizo

+++ database/migrations/2019_07_05_164459_create_permiso_rol_table.php
@@ -14,15 +14,15 @@ class CreatePermisoRolTable extends Migration
     public function up()
     {
         Schema::create('permiso_rol', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->boolean('estado');
 
-            $table->unsignedInteger('rol_id');
+            $table->unsignedBigInteger('rol_id');
             
             $table->foreign('rol_id','fk_permisorol_rol')->references('id')->on('rol')->onDelete('restrict')->onUpdate('restrict');
 
-            $table->unsignedInteger('permiso_id');
+            $table->unsignedBigInteger('permiso_id');
 
             $table->foreign('permiso_id','fk_permisorol_permiso')->references('id')->on('permiso')->onDelete('restrict')->onUpdate('restrict');
