+++ database/migrations/2019_07_05_165429_create_libro_prestamo_table.php
@@ -14,12 +14,12 @@ class CreateLibroPrestamoTable extends Migration
     public function up()
     {
         Schema::create('libro_prestamo', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
 
-            $table->unsignedInteger('usuario_id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id','fk_libroprestamo_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
 
-            $table->unsignedInteger('libro_id');
+            $table->unsignedBigInteger('libro_id');
             $table->foreign('libro_id','fk_libroprestamo_libro')->references('id')->on('libro')->onDelete('restrict')->onUpdate('restrict');
 
             $table->date('fecha_prestamo');
