+++ database/migrations/2019_07_13_162557_create_menu_table.php
@@ -14,8 +14,8 @@ class CreateMenuTable extends Migration
     public function up()
     {
         Schema::create('menu', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('menu_id')->default(0);
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('menu_id')->default(0);
             $table->string('nombre',50);
             $table->string('url',100);
             $table->unsignedInteger('orden')->default(0);
